# Conflicts:
#	cas-server-documentation/installation/Ticket-Registry-Replication-Encryption.md
#	cas-server-documentation/installation/Webflow-Customization.md
#	cas-server-extension-clearpass/src/main/java/org/jasig/cas/extension/clearpass/TicketRegistryDecorator.java
#	cas-server-webapp-config/src/main/java/org/jasig/cas/config/CasAuditTrailConfiguration.java
#	cas-server-webapp-config/src/main/java/org/jasig/cas/config/CasSecurityContextConfiguration.java
#	cas-server-webapp-config/src/main/java/org/jasig/cas/config/CasWebAppConfiguration.java
#	cas-server-webapp-config/src/main/java/org/jasig/cas/config/CasWebflowContextConfiguration.java
#	cas-server-webapp/src/main/webapp/WEB-INF/cas-servlet.xml
#	cas-server-webapp/src/main/webapp/WEB-INF/spring-configuration/applicationContext.xml

+++ cas-server-core-tickets/src/main/java/org/jasig/cas/ticket/registry/DefaultTicketRegistry.java
@@ -88,12 +88,12 @@ public final class DefaultTicketRegistry extends AbstractTicketRegistry {
     }
 
     @Override
-    public int sessionCount() {
+    public long sessionCount() {
         return (int) this.cache.values().stream().filter(t -> t instanceof TicketGrantingTicket).count();
     }
 
     @Override
-    public int serviceTicketCount() {
+    public long serviceTicketCount() {
         return (int) this.cache.values().stream().filter(t -> t instanceof ServiceTicket).count();
     }
